Got a alpha version out!
I'll have Kennan look over the app and tell me what he thinks. I
think it will hit a fair bit of his initial requirements. But there
is much to be improved. Firstly, and most obviously to him, will be
the fact that it is designed for a tablet not a phone. So I'll need
to reconfigure the layout to get it to work better on a smaller
display. The ability to remove drinks would also be valuable, and
edit them. In case you enter some wrong info or remembered another
drink and want to add it later.