UI updates mostly.
I fixed the negative BAC. Not only was that being kinda silly, that app could tell you that you had a BAC of zero or less immediately after drinking your n>1 drink. So it screwed with the accuracy. I just had it bottom out. So if a drink says it's going to give you < 0 BAC, it gives you 0.

I moved the BAC total to the bottom and added a naming field on the main screen. I'll look into the momentum on the ABV and OZ selectors. I think it can be done, but just didn't get to it tonight. The other thought I had was to give a dial to each digit. Which I'll do if I can't give it momentum.

The drink list now comes down with an animation. It's a bit dramatic and silly, but I so I can tone it down. I just felt like doing something fun. I believe I have fixed the squishing issue. And if I haven't fixed it sufficiently I can easily up the size of the drinks listed. I darkened the drink list background. Drinks now have an edit button and a delete button. Delete works, edit doesn't. I got stuck on some formatting issues or I might have been able to get edit to work.