Added post-mortem.
I'm done with this project. It has run it's course and I have found
it to be not possible, or at least not worth while. So I have
writen a new readme that includes a post-mortem on things I learned
and reasons why it didn't work. I also added the pictures into the
readme.

It was a good run and I learned quite a bit. Rest in peace
Flight Night.